Convert youtube-player to TypeScript
